fix(search-field-clear): remove style from primitive
