--- packages/ng-primitives/search-field-clear/README.md
@@ -1,3 +1,0 @@
-# ng-primitives/search-field-clear
-
-Secondary entry point of `ng-primitives`. It can be used by importing from `ng-primitives/search-field-clear`.
