--- packages/ng-primitives/search-field-clear/ng-package.json
@@ -1,5 +1,0 @@
-{
-  "lib": {
-    "entryFile": "src/index.ts"
-  }
-}
