--- packages/ng-primitives/search-field-clear/src/index.ts
@@ -1,9 +1,0 @@
-/**
- * Copyright © 2024 Angular Primitives.
- * https://github.com/ng-primitives/ng-primitives
- *
- * This source code is licensed under the Apache 2.0 license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-export { NgpSearchFieldClear } from './search-field-clear/search-field-clear.directive';
