+++ packages/ng-primitives/search/src/search-field-clear/search-field-clear.directive.spec.ts
