+++ tsconfig.base.json
@@ -34,9 +34,6 @@
       "ng-primitives/resize": ["packages/ng-primitives/resize/src/index.ts"],
       "ng-primitives/roving-focus": ["packages/ng-primitives/roving-focus/src/index.ts"],
       "ng-primitives/search": ["packages/ng-primitives/search/src/index.ts"],
-      "ng-primitives/search-field-clear": [
-        "packages/ng-primitives/search-field-clear/src/index.ts"
-      ],
       "ng-primitives/select": ["packages/ng-primitives/select/src/index.ts"],
       "ng-primitives/slider": ["packages/ng-primitives/slider/src/index.ts"],
       "ng-primitives/switch": ["packages/ng-primitives/switch/src/index.ts"],
